back/requirements.txt
appnope==0.1.0
astroid==2.3.2
autopep8==1.4.4
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
coverage==4.5.4
coveralls==1.8.2
decorator==4.4.0
Django==2.2.5
django-rest-framework==0.1.0
djangorestframework==3.10.3
docopt==0.6.2
entrypoints==0.3
flake8==3.7.8
flake8-django==0.0.4
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
lazy-object-proxy==1.4.2
mccabe==0.6.1
olefile==0.46
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.4.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
wrapt==1.11.2
